remove dead code and collapse file structure
